+++ outputs.tf
@@ -1,15 +1,20 @@
-output "stage_execution_arn" {
-  value = aws_api_gateway_stage.stage.execution_arn
+locals {
+  redeployment_triggers = merge(
+    {
+      for path, definition in local.resource_definitions:
+        "resource-${path}" => aws_api_gateway_resource.resource[path].id
+    },
+    {
+      for key, definition in local.method_definitions:
+        "method-${key}" => aws_api_gateway_method.method[key].id
+    },
+    {
+      for key, definition in local.integration_definitions:
+        "integration-${key}" => aws_api_gateway_integration.integration[key].id
+    }
+  )
 }
 
-output "stage_invoke_url" {
-  value = aws_api_gateway_stage.stage.invoke_url
-}
-
-output "stage_name" {
-  value = aws_api_gateway_stage.stage.stage_name
-}
-
-output "lambda_role_arn" {
-  value = aws_iam_role.proxy_lambda_execution_role.arn
+output "api_gateway_redeployment_triggers" {
+  value = local.redeployment_triggers
 }