+++ spec/unit/infra/prerequisites/main.tf
@@ -1,22 +1,26 @@
-module "base_network" {
-  source  = "infrablocks/base-networking/aws"
-  version = "4.0.0"
+module "api_gateway" {
+  source  = "infrablocks/api-gateway/aws"
+  version = "2.0.0-rc.2"
 
   region = var.region
-  vpc_cidr = var.vpc_cidr
-  availability_zones = var.availability_zones
 
   component = var.component
   deployment_identifier = var.deployment_identifier
-
-  private_zone_id = var.private_zone_id
 }
 
-resource "aws_api_gateway_rest_api" "api" {
-  name        = "${var.component}-${var.deployment_identifier}"
-  description = "${var.component}-${var.deployment_identifier} REST API"
+module "lambda" {
+  source  = "infrablocks/lambda/aws"
+  version = "2.0.0-rc.5"
 
-  endpoint_configuration {
-    types = ["REGIONAL"]
-  }
+  region = var.region
+
+  component             = var.component
+  deployment_identifier = var.deployment_identifier
+
+  lambda_function_name = "lambda-${var.deployment_identifier}"
+  lambda_description   = "test lambda"
+
+  lambda_zip_path = "${path.root}/lambda.zip"
+  lambda_handler  = "handler.hello"
+  lambda_runtime  = "nodejs16.x"
 }