+++ spec/unit/infra/prerequisites/outputs.tf
@@ -1,23 +1,11 @@
-output "vpc_id" {
-  value = module.base_network.vpc_id
+output "api_gateway_rest_api_id" {
+  value = module.api_gateway.api_gateway_rest_api_id
 }
 
-output "vpc_cidr" {
-  value = module.base_network.vpc_cidr
+output "api_gateway_rest_api_root_resource_id" {
+  value = module.api_gateway.api_gateway_rest_api_root_resource_id
 }
 
-output "public_subnet_ids" {
-  value = module.base_network.public_subnet_ids
-}
-
-output "private_subnet_ids" {
-  value = module.base_network.private_subnet_ids
-}
-
-output "api_gateway_id" {
-  value = aws_api_gateway_rest_api.api.id
-}
-
-output "api_gateway_root_resource_id" {
-  value = aws_api_gateway_rest_api.api.root_resource_id
+output "lambda_function_name" {
+  value = module.lambda.lambda_function_name
 }